Remove duplicate FileUtils methods
