+++ smack-core/src/main/java/org/jivesoftware/smack/util/FileUtils.java
@@ -31,7 +31,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import java.util.Stack;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -154,50 +153,6 @@ public final class FileUtils {
         catch (IOException e) {
             LOGGER.log(Level.WARNING, "writeFile", e);
             return false;
-        }
-    }
-
-    public static void deleteDirectory(File root) {
-        if (!root.exists()) {
-            return;
-        }
-        File[] currList;
-        Stack<File> stack = new Stack<>();
-        stack.push(root);
-        while (!stack.isEmpty()) {
-            if (stack.lastElement().isDirectory()) {
-                currList = stack.lastElement().listFiles();
-                if (currList != null && currList.length > 0) {
-                    for (File curr : currList) {
-                        stack.push(curr);
-                    }
-                } else {
-                    stack.pop().delete();
-                }
-            } else {
-                stack.pop().delete();
-            }
-        }
-    }
-
-    /**
-     * Returns a {@link File} pointing to a temporary directory. On unix like systems this might be {@code /tmp}
-     * for example.
-     * If {@code suffix} is not null, the returned file points to {@code <temp>/suffix}.
-     *
-     * @param suffix optional path suffix
-     * @return temp directory
-     */
-    public static File getTempDir(String suffix) {
-        String temp = System.getProperty("java.io.tmpdir");
-        if (temp == null) {
-            temp = "tmp";
-        }
-
-        if (suffix == null) {
-            return new File(temp);
-        } else {
-            return new File(temp, suffix);
         }
     }
 }
