+++ smack-integration-test/src/main/java/org/jivesoftware/smackx/ox/OXSecretKeyBackupIntegrationTest.java
@@ -32,7 +32,7 @@ import java.util.logging.Level;
 
 import org.jivesoftware.smack.SmackException;
 import org.jivesoftware.smack.XMPPException;
-import org.jivesoftware.smack.util.FileUtils;
+import org.jivesoftware.smack.test.util.FileTestUtil;
 import org.jivesoftware.smack.util.StringUtils;
 import org.jivesoftware.smackx.ox.callback.backup.AskForBackupCodeCallback;
 import org.jivesoftware.smackx.ox.callback.backup.DisplayBackupCodeCallback;
@@ -63,8 +63,8 @@ import org.pgpainless.key.protection.UnprotectedKeysProtector;
 public class OXSecretKeyBackupIntegrationTest extends AbstractOpenPgpIntegrationTest {
 
     private static final String sessionId = StringUtils.randomString(10);
-    private static final File beforePath = FileUtils.getTempDir("ox_backup_" + sessionId);
-    private static final File afterPath = FileUtils.getTempDir("ox_restore_" + sessionId);
+    private static final File beforePath = FileTestUtil.getTempDir("ox_backup_" + sessionId);
+    private static final File afterPath = FileTestUtil.getTempDir("ox_restore_" + sessionId);
 
     private String backupCode = null;
 
@@ -110,8 +110,8 @@ public class OXSecretKeyBackupIntegrationTest extends AbstractOpenPgpIntegration
     @BeforeClass
     public static void cleanStore() {
         LOGGER.log(Level.INFO, "Delete store directories...");
-        FileUtils.deleteDirectory(afterPath);
-        FileUtils.deleteDirectory(beforePath);
+        FileTestUtil.deleteDirectory(afterPath);
+        FileTestUtil.deleteDirectory(beforePath);
     }
 
     @After
