+++ smack-openpgp/src/test/java/org/jivesoftware/smackx/ox/SecretKeyBackupHelperTest.java
@@ -28,8 +28,8 @@ import java.util.Arrays;
 import java.util.Collections;
 
 import org.jivesoftware.smack.DummyConnection;
+import org.jivesoftware.smack.test.util.FileTestUtil;
 import org.jivesoftware.smack.test.util.SmackTestSuite;
-import org.jivesoftware.smack.util.FileUtils;
 import org.jivesoftware.smackx.ox.crypto.PainlessOpenPgpProvider;
 import org.jivesoftware.smackx.ox.element.SecretkeyElement;
 import org.jivesoftware.smackx.ox.exception.InvalidBackupCodeException;
@@ -54,7 +54,7 @@ public class SecretKeyBackupHelperTest extends SmackTestSuite {
     private static final File basePath;
 
     static {
-        basePath = FileUtils.getTempDir("ox_secret_keys");
+        basePath = FileTestUtil.getTempDir("ox_secret_keys");
     }
 
     @Test
@@ -98,6 +98,6 @@ public class SecretKeyBackupHelperTest extends SmackTestSuite {
     @AfterClass
     @BeforeClass
     public static void deleteDirs() {
-        FileUtils.deleteDirectory(basePath);
+        FileTestUtil.deleteDirectory(basePath);
     }
 }
